Added security part

+++ src/main/java/com/mindfire/rentingit/services/AddProducts.java
@@ -3,6 +3,9 @@
  */
 package com.mindfire.rentingit.services;
 
+import java.util.List;
+import java.util.Optional;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
@@ -12,14 +15,20 @@ import com.mindfire.rentingit.dto.request.ProductRegisterRequest;
 
 import com.mindfire.rentingit.dto.response.MessageResponse;
 import com.mindfire.rentingit.entity.Product;
+import com.mindfire.rentingit.entity.User;
 import com.mindfire.rentingit.exception.ResourceNotFoundException;
 import com.mindfire.rentingit.repository.ProductRepository;
+import com.mindfire.rentingit.repository.UserRepository;
 
 @Service
 public class AddProducts {
 
 	@Autowired
 	ProductRepository productRepository;
+	@Autowired
+	UserRepository userRepository;
+	@Autowired
+	CustomUserDetailService customUserDetails;
 	@Autowired
 	Message msg;
 
@@ -65,12 +74,58 @@ public class AddProducts {
 
 	}
 
-	// method for deleting the registered products based upon id
-	public ResponseEntity<Product> deleteProductsById(long productId) {
+	// method for deleting the registered products based upon id by user
+	public ResponseEntity<?> deleteProductsById(long productId) {
+		// checking for the request if it is made by the current loggedIn user or not
+
 		Product existingProduct = this.productRepository.findById(productId)
 				.orElseThrow(() -> new ResourceNotFoundException("Product not found with this id :" + productId));
-		this.productRepository.delete(existingProduct);
-		return ResponseEntity.ok().build();
+
+		//getting the ownerID of product
+		long currentOwnerId = existingProduct.getOwnerId();
+		
+
+		User currentUser = customUserDetails.getCurrentUser();
+		long currentUserId = currentUser.getId();
+
+		boolean flag;
+		if (currentUserId == currentOwnerId ) {
+			flag = true;
+
+		} else {
+			flag = false;
+		}
+
+		if (flag == true) {
+			this.productRepository.delete(existingProduct);
+			return ResponseEntity.ok(new MessageResponse(msg.PRODUCTS_DELETED));
+		} else {
+			return ResponseEntity.ok(new MessageResponse(msg.CANT_DELETE_PROD));
+		}
 	}
 
+	// deleting the products by id for admin
+	public ResponseEntity<?> deleteProducts(long productId) {
+		// checking for the request if it is made by the current loggedIn user or not
+
+		Product existingProduct = this.productRepository.findById(productId)
+				.orElseThrow(() -> new ResourceNotFoundException("Product not found with this id :" + productId));
+
+		this.productRepository.delete(existingProduct);
+		return ResponseEntity.ok(new MessageResponse(msg.PRODUCTS_DELETED));
+
+	}
+	
+	//getting all products of currently loggedin user
+	public List<Product> findProductsOfUser(){
+		
+		User currentUser = customUserDetails.getCurrentUser();
+		long userId = currentUser.getId();
+		
+		int ownerID = (int)userId;
+		
+		List<Product> p = this.productRepository.findByOwnerId(ownerID);
+		
+		return p;
+	}
 }
